refactor(app): modularize API client into per-domain modules (#989)
Split lib/api.ts into lib/api/{client,auth,jobs,payments}.ts, extracting
a shared request wrapper and the auth/jobs/payments endpoint groups.
Updates call sites to import from the new modules.
