+++ packages/app/src/app/[locale]/jobs/new/page.tsx
@@ -5,7 +5,8 @@ import { useRouter } from "next/navigation";
 import Link from "next/link";
 import { ArrowLeft, Loader2, X, Plus } from "lucide-react";
 import { useAuth } from "@/context/AuthContext";
-import { createJob, getCategories } from "@/lib/api";
+import { getCategories } from "@/lib/api";
+import { createJob } from "@/lib/api/jobs";
 import type { Category } from "@/types";
 
 export default function NewJobPage() {
